Update README.md
I am a non-resident fellow at the Quincy Institute and a research fellow at the Center for Global Development and Sustainability at Brandeis University. As a welfare economist, he served as a research fellow at the Crown Center for Middle East Studies from 2017 to 2025, where he studied the political economy of U.S.–Iran policy and the social welfare effects of economic sanctions across Iran and the broader Middle East and North Africa.